refactor: split app error facade internals
